Adds scalar support to XML deserialization.
This change reads the input string from the Reader in the IOContext supplied to the FromXmlParser to try to determine if the input XML represents a scalar value.  If so, when asking for the nextToken(), this change skips the tokens that don't pertain to scalars to get to the scalar value contained inside the XML.  This shouldn't impact other XML strings.