Create reproducer test case for #509
This adds two additional test cases. Those tests attempt to deserialize
the two classes Data and MetaData.

The MetaData class simply contains a list of Data objects. The Data
class contains a "key", which is derived from the attribute of the XML
node, and a "content" list, which is the arbitrary data stored inside
that node.